requirements.txt
# TeleTask Bot Dependencies
# Generated by install.sh

# Telegram Bot Framework
python-telegram-bot>=21.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Migrations
alembic>=1.13.0

# Scheduler
APScheduler>=3.10.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
aiohttp>=3.9.0

# Timezone
pytz>=2024.1

# Date utilities
python-dateutil>=2.8.0

# Cache (optional)
redis>=5.0.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Google Calendar (optional)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Export/Report generation
openpyxl>=3.1.0
matplotlib>=3.8.0
reportlab>=4.0.0

# System monitoring
psutil>=5.9.0

# Security - Token encryption
cryptography>=42.0.0
